Move cloud onboarding tests to the cloud integration
